requirements.txt
asgiref==3.5.0
boto3==1.20.54
botocore==1.23.54
certifi==2021.10.8
charset-normalizer==2.0.12
Django
django-environ==0.8.1
django-storages==1.12.3
djangorestframework
idna==3.3
jmespath==0.10.0
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
s3transfer==0.5.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8

document
Pillow